드래그 자율화 필요
